Port product_price_history to 12
