+++ product_price_history/__init__.py
@@ -1,4 +1,1 @@
-# Copyright 2020 Andrea Piovesana (Associazione PNLUG - Gruppo Odoo)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from . import models
