+++ product_price_history/__manifest__.py
@@ -2,15 +2,16 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
-    'name': "Product Price History View",
-    'summary': "Product Price History View",
-    'description': "Product Price History View",
-    'version': '11.0.1.0.0',
+    'name': "Product Cost Price History Views",
+    'summary': "Handle standard price history records",
+    'version': '12.0.1.0.0',
     'category': 'Product',
     'author': 'Pordenone Linux User Group (PNLUG), Odoo Community Association (OCA)',
+    'maintainers': ["marcelofrare", "andreampiovesana"],
     'website': 'https://github.com/OCA/product-attribute',
+    'license': 'AGPL-3',
     'depends': [
-        'product',
+        'stock',
     ],
     'data': [
         'views/product_price_history_view.xml',
