--- product_price_history/i18n/product_price_history.pot
@@ -1,68 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* product_price_history
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 11.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-07-05 15:14+0000\n"
-"PO-Revision-Date: 2020-07-05 15:14+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: product_price_history
-#: model:ir.ui.view,arch_db:product_price_history.view_product_price_history_filter
-msgid "Date"
-msgstr ""
-
-#. module: product_price_history
-#: model:ir.ui.view,arch_db:product_price_history.view_product_price_history_filter
-msgid "Group By"
-msgstr ""
-
-#. module: product_price_history
-#: model:ir.actions.act_window,name:product_price_history.action_product_price_history_view
-#: model:ir.actions.act_window,name:product_price_history.action_product_template_only_form_price_history
-#: model:ir.ui.view,arch_db:product_price_history.view_product_normal_form_view_price_history
-#: model:ir.ui.view,arch_db:product_price_history.view_product_template_only_form_price_history
-msgid "Price History"
-msgstr ""
-
-#. module: product_price_history
-#: model:ir.model.fields,field_description:product_price_history.field_product_product_price_history_count
-#: model:ir.model.fields,field_description:product_price_history.field_product_template_price_history_count
-msgid "Price History Count"
-msgstr ""
-
-#. module: product_price_history
-#: model:ir.model,name:product_price_history.model_product_product
-#: model:ir.ui.view,arch_db:product_price_history.view_product_price_history_filter
-msgid "Product"
-msgstr ""
-
-#. module: product_price_history
-#: model:ir.ui.menu,name:product_price_history.menu_product_price_history
-msgid "Product Price History"
-msgstr ""
-
-#. module: product_price_history
-#: model:ir.model,name:product_price_history.model_product_template
-#: model:ir.model.fields,field_description:product_price_history.field_product_price_history_product_tmpl_id
-msgid "Product Template"
-msgstr ""
-
-#. module: product_price_history
-#: model:ir.ui.view,arch_db:product_price_history.view_product_price_history_filter
-msgid "Search Product Price History"
-msgstr ""
-
-#. module: product_price_history
-#: model:ir.model,name:product_price_history.model_product_price_history
-msgid "product.price.history"
-msgstr ""
-
