+++ product_price_history/models/__init__.py
@@ -1,3 +1,3 @@
-# Copyright 2020 Andrea Piovesana (Associazione PNLUG - Gruppo Odoo)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-from . import product
+from . import product_product
+from . import product_template
+from . import product_price_history
