--- product_price_history/readme/CONFIGURE.rst
@@ -1,1 +1,0 @@
-These module don't need configuration
