+++ product_price_history/readme/CONTRIBUTORS.rst
@@ -1,2 +1,6 @@
 * Andrea Piovesana <andrea.m.piovesana@gmail.com>
 * Marcelo Frare <mf2965@gmail.com>
+
+* `TAKOBI <https://takobi.online>`_:
+
+  * Lorenzo Battistini
