--- product_price_history/readme/INSTALL.rst
@@ -1,1 +1,0 @@
-Install with standard method
