--- product_price_history/readme/ROADMAP.rst
@@ -1,1 +1,0 @@
-*
